Removing client callback retries mechanism, that tried to reach unreachable clients X times before giving up.  Now first unsuccessful attempt to reach a client will cause client to be unregistered for that event.
Removing exception stack from ApiConnectionError class, as it was too much noise on the receiving end, with very little value.  Also removing message for this error, as it also adds no value, only noise, and key elements (unreachable url and body sent) are already in the payload.

Updating Registration class to automatically unregister with EventDispatch if its client was unreachable.

Updating test cases for these changes.